pg_managementPRD/requirements.txt
##✅ RECOMMENDED
Django>=6.0,<6.1  #Core framework
djangorestframework>=3.15 #You are using DRF for API development
djangorestframework-simplejwt>=5.3 #JWT authentication
django-cors-headers>=4.3 #CORS middleware(handling)

#🧪 OPTIONAL (ONLY IF USED ELSEWHERE)
python-dotenv
psycopg2-binary
Pillow

# 🛠️ HOW TO INSTALL
pip install -r requirements.txt

# ✅ VERIFY INSTALLATION
python -m django --version

pip show djangorestframework
pip show django-cors-headers
pip show djangorestframework-simplejwt


python-dotenv #if used



